+++ rychkov.mihail/P2/taylors_series.cpp
@@ -1,18 +1,17 @@
-#include "taylors_row.hpp"
+#include "taylors_series.hpp"
 
 #include <stdexcept>
 #include <cmath>
 
 double rychkov::formulas::calcUnoDivSqr(double x, size_t maxDepth, double absError)
 {
-  double xSqr = x * x;
   double lastSummand = 1.0;
   double result = 1.0;
   for (size_t i = 0; i < maxDepth; i++)
   {
     lastSummand *= (i * 2) + 1;
     lastSummand /= (i * 2) + 2;
-    lastSummand *= xSqr;
+    lastSummand *= x * x;
 
     if (lastSummand < absError)
     {
